bgp manager,speaker: avoid duplicate event processing
Prior to this commit it was possible for the BGP advertisement business
logic to replay stale events.

This scenario would happen when concurrently enqueing new events while
requeing events which failed to be processed.

The requeued event has the ability to get "replayed" after a
concurrently enqeued new event, leading to an ABA problem.

This commit adds a new type called "fence.Fencer" which keeps track of
the largest revision number seen for a particular UUID.

Both revision numbers and UUID's are kubernetes (etcd) values crafted
for this specific purpose.

Signed-off-by: Louis DeLosSantos <louis.delos@isovalent.com>