bpf: snat_v4_needed: single egress gw policy check
This patch refactors snat_v4_needed() to perform a single check for the
egress gateway. Previous code performed two checks, one of which was
outside of `if (!info)`, where `info` was the result of an ipcache
lookup on the destination.

Since the ip cache lookup will always return something, we can merge the
two checks into one to simplify the code.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>